Repository hierarchy reconstructed
